+++ unit_tests/extractor/graph_compressor.cpp
@@ -68,7 +68,6 @@ BOOST_AUTO_TEST_CASE(long_road_test)
     std::unordered_set<NodeID> barrier_nodes;
     std::unordered_set<NodeID> traffic_lights;
     std::vector<TurnRestriction> restrictions;
-    std::vector<ConditionalTurnRestriction> conditional_restrictions;
     std::vector<NodeBasedEdgeAnnotation> annotations(1);
     CompressedEdgeContainer container;
     test::MockScriptingEnvironment scripting_environment;
@@ -92,7 +91,6 @@ BOOST_AUTO_TEST_CASE(long_road_test)
                         traffic_lights,
                         scripting_environment,
                         restrictions,
-                        conditional_restrictions,
                         maneuver_overrides,
                         graph,
                         annotations,
@@ -116,7 +114,6 @@ BOOST_AUTO_TEST_CASE(loop_test)
     std::unordered_set<NodeID> barrier_nodes;
     std::unordered_set<NodeID> traffic_lights;
     std::vector<TurnRestriction> restrictions;
-    std::vector<ConditionalTurnRestriction> conditional_restrictions;
     CompressedEdgeContainer container;
     std::vector<NodeBasedEdgeAnnotation> annotations(1);
     test::MockScriptingEnvironment scripting_environment;
@@ -154,7 +151,6 @@ BOOST_AUTO_TEST_CASE(loop_test)
                         traffic_lights,
                         scripting_environment,
                         restrictions,
-                        conditional_restrictions,
                         maneuver_overrides,
                         graph,
                         annotations,
@@ -182,7 +178,6 @@ BOOST_AUTO_TEST_CASE(t_intersection)
     std::unordered_set<NodeID> traffic_lights;
     std::vector<NodeBasedEdgeAnnotation> annotations(1);
     std::vector<TurnRestriction> restrictions;
-    std::vector<ConditionalTurnRestriction> conditional_restrictions;
     CompressedEdgeContainer container;
     test::MockScriptingEnvironment scripting_environment;
     std::vector<UnresolvedManeuverOverride> maneuver_overrides;
@@ -205,7 +200,6 @@ BOOST_AUTO_TEST_CASE(t_intersection)
                         traffic_lights,
                         scripting_environment,
                         restrictions,
-                        conditional_restrictions,
                         maneuver_overrides,
                         graph,
                         annotations,
@@ -227,7 +221,6 @@ BOOST_AUTO_TEST_CASE(street_name_changes)
     std::unordered_set<NodeID> traffic_lights;
     std::vector<NodeBasedEdgeAnnotation> annotations(2);
     std::vector<TurnRestriction> restrictions;
-    std::vector<ConditionalTurnRestriction> conditional_restrictions;
     CompressedEdgeContainer container;
     test::MockScriptingEnvironment scripting_environment;
     std::vector<UnresolvedManeuverOverride> maneuver_overrides;
@@ -246,7 +239,6 @@ BOOST_AUTO_TEST_CASE(street_name_changes)
                         traffic_lights,
                         scripting_environment,
                         restrictions,
-                        conditional_restrictions,
                         maneuver_overrides,
                         graph,
                         annotations,
@@ -267,7 +259,6 @@ BOOST_AUTO_TEST_CASE(direction_changes)
     std::unordered_set<NodeID> traffic_lights;
     std::vector<NodeBasedEdgeAnnotation> annotations(1);
     std::vector<TurnRestriction> restrictions;
-    std::vector<ConditionalTurnRestriction> conditional_restrictions;
     CompressedEdgeContainer container;
     test::MockScriptingEnvironment scripting_environment;
     std::vector<UnresolvedManeuverOverride> maneuver_overrides;
@@ -282,7 +273,6 @@ BOOST_AUTO_TEST_CASE(direction_changes)
                         traffic_lights,
                         scripting_environment,
                         restrictions,
-                        conditional_restrictions,
                         maneuver_overrides,
                         graph,
                         annotations,