+++ unit_tests/extractor/intersection_analysis_tests.cpp
@@ -24,7 +24,6 @@ BOOST_AUTO_TEST_CASE(simple_intersection_connectivity)
         {EMPTY_NAMEID, 0, INAVLID_CLASS_DATA, TRAVEL_MODE_DRIVING, false},
         {EMPTY_NAMEID, 1, INAVLID_CLASS_DATA, TRAVEL_MODE_DRIVING, false}};
     std::vector<TurnRestriction> restrictions{TurnRestriction{NodeRestriction{0, 2, 1}, false}};
-    std::vector<ConditionalTurnRestriction> conditional_restrictions;
     CompressedEdgeContainer container;
     test::MockScriptingEnvironment scripting_environment;
     std::vector<UnresolvedManeuverOverride> maneuver_overrides;
@@ -90,7 +89,6 @@ BOOST_AUTO_TEST_CASE(simple_intersection_connectivity)
                                traffic_lights,
                                scripting_environment,
                                restrictions,
-                               conditional_restrictions,
                                maneuver_overrides,
                                graph,
                                annotations,
@@ -101,7 +99,8 @@ BOOST_AUTO_TEST_CASE(simple_intersection_connectivity)
 
     EdgeBasedNodeDataContainer node_data_container(
         std::vector<EdgeBasedNode>(graph.GetNumberOfEdges()), annotations);
-    RestrictionMap restriction_map(restrictions, IndexNodeByFromAndVia());
+    RestrictionGraph restriction_graph = constructRestrictionGraph(restrictions);
+    RestrictionMap restriction_map(restriction_graph);
 
     const auto connectivity_matrix = [&](NodeID node) {
         std::vector<bool> result;
@@ -156,7 +155,6 @@ BOOST_AUTO_TEST_CASE(roundabout_intersection_connectivity)
     std::unordered_set<NodeID> traffic_lights;
     std::vector<NodeBasedEdgeAnnotation> annotations;
     std::vector<TurnRestriction> restrictions;
-    std::vector<ConditionalTurnRestriction> conditional_restrictions;
     CompressedEdgeContainer container;
     test::MockScriptingEnvironment scripting_environment;
     std::vector<UnresolvedManeuverOverride> maneuver_overrides;
@@ -214,7 +212,6 @@ BOOST_AUTO_TEST_CASE(roundabout_intersection_connectivity)
                                traffic_lights,
                                scripting_environment,
                                restrictions,
-                               conditional_restrictions,
                                maneuver_overrides,
                                graph,
                                annotations,
@@ -225,7 +222,9 @@ BOOST_AUTO_TEST_CASE(roundabout_intersection_connectivity)
 
     EdgeBasedNodeDataContainer node_data_container(
         std::vector<EdgeBasedNode>(graph.GetNumberOfEdges()), annotations);
-    RestrictionMap restriction_map(restrictions, IndexNodeByFromAndVia());
+
+    RestrictionGraph restriction_graph = constructRestrictionGraph(restrictions);
+    RestrictionMap restriction_map(restriction_graph);
 
     const auto connectivity_matrix = [&](NodeID node) {
         std::vector<bool> result;
@@ -263,7 +262,6 @@ BOOST_AUTO_TEST_CASE(skip_degree_two_nodes)
     std::unordered_set<NodeID> traffic_lights{2};
     std::vector<NodeBasedEdgeAnnotation> annotations(1);
     std::vector<TurnRestriction> restrictions;
-    std::vector<ConditionalTurnRestriction> conditional_restrictions;
     CompressedEdgeContainer container;
     test::MockScriptingEnvironment scripting_environment;
     std::vector<UnresolvedManeuverOverride> maneuver_overrides;
@@ -306,7 +304,6 @@ BOOST_AUTO_TEST_CASE(skip_degree_two_nodes)
                                traffic_lights,
                                scripting_environment,
                                restrictions,
-                               conditional_restrictions,
                                maneuver_overrides,
                                graph,
                                annotations,